requirements/dev.txt
# This file pulls in everything a developer needs. If it's a basic package
# needed to run the site, it belongs in requirements/prod.txt. If it's a
# package for developers (testing, docs, etc.), it goes in this file.

-r prod.txt

# Documentation
Sphinx==1.1

# Testing
nose==1.0.0
mock==0.8.0
-e git://github.com/jbalogh/django-nose.git#egg=django_nose
-e git://github.com/jbalogh/test-utils.git#egg=test-utils
lxml==2.1
pyquery

# L10n
translate-toolkit==1.8.0

# Debug tools
django-extensions
Werkzeug

coverage
nosenicedots
factory-boy
